# raspyxauth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for authentication/authorization API service",
        "title": "JWT authentication/authorization service",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/auth/api",
    "paths": {
...
        "/v1/users/{uuid}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updating user information by their UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Updating user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data has been successfully updated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request format",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "404": {
                        "description": "User was not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "409": {
                        "description": "Conflict when updating user",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deleting a user by their UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Deleting user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully deleted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request format",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "404": {
                        "description": "User was not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "409": {
                        "description": "Conflict when deleting user",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseAPI": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/models.ErrorResponse"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 50
                },
                "username": {
                    "type": "string",
                    "example": "username"
                }
            }
        },
...
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string",
                    "example": "4b6c34bd-01f0-4fbe-be65-217984b3e33d"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}